+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * [FIX] Do not navigate to currently open profile register #244
 * [FIX] right panel should display refresh content after changes to node #249
-* [FIX] tree should be updated afgter destroy date has been set #249
+* [FIX] tree should be updated after destroy date has been set #249
+* [FIX] Open deleted nodes via trash menu does not list any children #254
 
 
 ## 3.2.5